fix: PostgreSQL, copy-on-write, and WebDAV production bugs from Phase 8 validation
Why: Running the suite on PostgreSQL and E2E in both modes exposed real bugs in
dual-backend stores and in the S3 copy-on-write / WebDAV blob paths.

What:
- fileNodesStore.insertAncestorRows: group PG multi-row VALUES placeholders
  (was flat $1..$N -> 42601); permissionRequestStore: $4::BIGINT cast
  (was 42804).
- blobStorageService.ensureExclusiveBlob: use upsertObjectMap (per-node
  orphan + next version) instead of orphanObject(s3_key)+insertObject(v1)
  which orphaned the original and 409'd on overwriting a copy (E2E-S3PG-004);
  corrected the unit test that encoded the wrong behavior.
- storage.js getPgPool: register pool.on('error') so a PG drop no longer
  crashes the process.
- WebDAV mode: recursive already-exists-tolerant MKCOL (utils/webdav
  ensureDirectoryExists), WebdavBlobStore.createDirectory,
  blobStorageService.createDirectoryWebdav (S3 no-op, orphaned_node
  fail-safe), wired into folders.js, userService.ensureUserHomeNode,
  cleanupService.ensureHomeOwnerAdminForAllUsers; folders.test.js asserts the
  MKCOL path.

Impact: full server suite green on SQLite + PostgreSQL (67/1136); WebDAV E2E
uploads now succeed (was 0%). Fail-loud orphaned_node contract preserved.

+++ server/domains/permissions/stores/permissionRequestStore.js
@@ -364,7 +364,7 @@ async function updateStatus(id, { status, resolvedBy } = {}) {
           `UPDATE permission_requests
               SET status = $2,
                   resolved_at = CASE WHEN $2 = $3 THEN NULL ELSE NOW() END,
-                  resolved_by = CASE WHEN $2 = $3 THEN NULL ELSE $4 END
+                  resolved_by = CASE WHEN $2 = $3 THEN NULL ELSE $4::BIGINT END
             WHERE id = $1
             RETURNING *`,
           [
